PR #951 in the wrong location
