+++ wrangles/format.py
@@ -121,6 +121,52 @@ def coalesce(input_list: list) -> list:
     Return the first not empty result for each row
     where each row has a list of possibilities
     """
+    if not input_list:
+        return []
+
+    # Use numpy optimization for rectangular (fixed-width) input (e.g. multi-column coalesce).
+    # Fall back to Python loop for ragged input (e.g. lists stored in a single column).
+    row_len = len(input_list[0])
+    if row_len > 0 and all(len(row) == row_len for row in input_list):
+        arr = _np.empty((len(input_list), row_len), dtype=object)
+        for i, row in enumerate(input_list):
+            for j, val in enumerate(row):
+                arr[i, j] = val if val is not None else ''
+
+        n_rows, m = arr.shape
+
+        # Fast C-level check: find first column per row where value != ''
+        mask = arr != ''
+        idx = _np.argmax(mask, axis=1)
+        has_any = mask.any(axis=1)
+        result = arr[_np.arange(n_rows), idx].copy()
+        result[~has_any] = ''
+
+        # arr != '' treats whitespace-only strings as non-empty; handle those edge cases
+        # by stripping only the selected values (1 per row) rather than the full matrix.
+        try:
+            stripped_result = _np.frompyfunc(str.strip, 1, 1)(result)
+        except (TypeError, AttributeError):
+            def _safe_strip(x):
+                try:
+                    return x.strip()
+                except AttributeError:
+                    return '' if not x else x
+            stripped_result = _np.frompyfunc(_safe_strip, 1, 1)(result)
+        needs_fallback = _np.where(has_any & (stripped_result == ''))[0]
+        for i in needs_fallback:
+            for j in range(int(idx[i]) + 1, m):
+                val = arr[i, j]
+                if isinstance(val, str):
+                    val = val.strip()
+                if val:
+                    result[i] = val
+                    break
+            else:
+                result[i] = ''
+
+        return result.tolist()
+
     output_list = []
     for row in input_list:
         output_row = ''
@@ -129,7 +175,6 @@ def coalesce(input_list: list) -> list:
             if value:
                 output_row = value
                 break
-
         output_list.append(output_row)
     return output_list
